+++ Src/Private/Images.ps1
@@ -1,0 +1,12 @@
+$script:Images = @{
+    'Microsoft_Logo' = 'Microsoft_Logo.png'
+    'ForestRoot' = 'Forrest_Root.png'
+    'AD_LOGO_Footer' = 'DMAD_Logo.png'
+    'AD_DC' = 'DomainController.png'
+    'AD_Domain' = 'AD_Domain.png'
+    'AD_Site' = 'Site.png'
+    'AD_Site_Subnet' = 'SubNet.png'
+    'AD_Site_Node' = 'SiteNode.png'
+    'AD_Certificate' = 'certificate.png'
+    'AD_PKI_Logo' = 'pkilogo.png'
+}